Improved search.php from jbsclm
"I've changed the order of the joins, search times have improved again.
The original 40+ seconds for the software query is down to 2.
As a different issue searching for a numeric, eg 5000 seems to hit on IP addresses after they have been through the ip_trans_to function, I'll try and understand what this is intendede to achieve.

The improved search.php is attached."

git-svn-id: svn://svn.code.sf.net/p/open-audit/code/trunk@1111 0fda6dae-eacc-4bf0-a90c-2bab3023e2ff